Use _Atomic types in ephemeron implementation & API
There is a small semantic change in that the atomic alterations to
ephemerons load/store a gc_ref instead of reaching into the gc_ref
to change its value. I think this is a bit nicer anyway, it makes
gc_ref more opaque.

atomic_load_explicit(...).value is probably legal, but clang fails
to compile it with "cannot compile this l-value expression yet" so
too bad.