attempted fix for nofl space overflow
i was seeing cases where object_granules was bigger than
limit_granules, which caused an overflow and broke stuff. This
makes it so that in such a case, we effectively treat the number
of granules as being the limit. But I don't understand this much
at all and I imagine there may be a deeper issue causing
object_granules to be that big in the first place.